requirements.txt
# Certificate Download Portal - Requirements
# Compatible with Python 3.7+

Flask>=2.0.0,<3.0.0
Flask-SQLAlchemy>=2.5.0,<4.0.0
Werkzeug>=2.0.0,<3.0.0
python-dotenv>=0.19.0
openpyxl>=3.0.0
pandas>=1.3.0,<2.0.0
PyMuPDF>=1.21.0
